requirement.txt
torch
torchvision
numpy
Pillow
tqdm
matplotlib
scikit-image
fastai